runtime: bounty: Refactor crate structure.
- move BountyActorManager to the separate actors.rs file.
- move BountyStageCalculator to the separate stage.rs file.
